deeplearn_eassy/requirements.txt
asgiref==3.2.3
aspy.yaml==1.3.0
baidu-aip==2.2.18.0
boto3==1.11.0
botocore==1.14.0
bottle==0.12.18
cachetools==4.0.0
certifi==2019.11.28
cfgv==2.0.1
chardet==3.0.4
Click==7.0
cma==2.7.0
colorama==0.4.3
colorlog==4.1.0
common==0.1.2
construct==2.5.3
cycler==0.10.0
data==0.4
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.2
django-cors-middleware==1.5.0
django-crispy-forms==1.8.1
django-formtools==2.2
django-haystack==2.8.1
django-import-export==2.0
django-redis==4.11.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
docutils==0.15.2
docx==0.2.4
drf-haystack==1.8.6
dual==0.0.8
dynamo3==0.4.10
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
flake8==3.7.9
Flask==1.1.1
flywheel==0.5.3
funcsigs==1.0.2
future==0.18.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
graphviz==0.13.2
grpcio==1.26.0
HTMLParser==0.0.2
httplib2==0.15.0
identify==1.4.9
idna==2.8
imageio==2.6.1
imageio-ffmpeg==0.3.0
importlib-metadata==1.4.0
itsdangerous==1.1.0
jdcal==1.4.1
jieba==0.42.1
Jinja2==2.10.3
jmespath==0.9.4
kiwisolver==1.1.0
lxml==4.4.2
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
more-itertools==8.1.0
moviepy==1.0.1
mysqlclient==1.4.6
nltk==3.4.5
nodeenv==1.3.4
oauthlib==3.1.0
objgraph==3.4.1
odfpy==1.4.0
openpyxl==3.0.2
packaging==20.0
paddle==1.0.2
paddlehub==1.4.1
paddlepaddle==1.6.3
pandas==0.25.3
pdfkit==0.6.1
peewee==3.13.1
pefile==2019.4.18
pluggy==0.13.1
pre-commit==1.21.0
prettytable==0.7.2
proglog==0.1.9
prox==0.0.17
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
PyDocX==0.9.10
pyflakes==2.1.1
PyJWT==1.7.1
pypandoc==1.4
pyparsing==2.4.6
pypiwin32==223
PySocks==1.7.1
python-dateutil==2.8.1
python-geoip-python3==1.3
python-ptrace==0.9.4
pytz==2019.3
pywin32==227
rarfile==3.1
redis==3.3.11
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.0
sentencepiece==0.1.85
tablib==0.14.0
tb-paddle==0.3.8
tight==0.1.0
toml==0.10.0
tox==3.14.3
translate==3.5.0
urllib3==1.25.7
virtualenv==16.7.9
visualdl==1.3.0
Whoosh==2.7.4
xlrd==1.2.0
xlwt==1.3.0
yapf==0.26.0
zipp==0.6.0
